+++ dyna_cli/components/screens/profile_select.py
@@ -1,0 +1,33 @@
+
+from textual.app import ComposeResult
+from textual.widgets import (
+    ListItem,
+    ListView,
+    Label,
+)
+from textual.screen import Screen
+from textual.message import Message
+from textual import log
+
+class ProfileSelectScreen(Screen):
+    BINDINGS = [("escape", "app.pop_screen", "Pop screen")]
+
+    class ProfileSelected(Message):
+        """"""
+
+        def __init__(self, profile: str) -> None:
+            self.profile = profile
+            super().__init__()
+
+    def compose(self) -> ComposeResult:
+        profiles = self.parent.profiles
+        yield ListView(
+            *[ListItem(Label(profile), id=profile) for profile in profiles],
+            id="profiles",
+        )
+
+    async def on_list_view_selected(self, selected) -> None:
+        self.post_message(self.ProfileSelected(selected.item.id))
+        self.app.pop_screen()
+
+
