+++ dyna_cli/components/screens/table_select.py
@@ -8,32 +8,10 @@ from textual.screen import Screen
 from textual.events import Key
 from textual.message import Message
 from textual.containers import Vertical
-from textual.widgets import Button, Input
-from dyna_cli.aws.session import get_all_regions
+from textual.widgets import  Input
 from dyna_cli.aws.ddb import get_ddb_client, list_all_tables
 from textual import log
-from botocore.exceptions import ClientError
 from textual.reactive import reactive
-
-
-class ErrorScreen(Screen):
-    BINDINGS = [("escape", "app.pop_screen", "Pop screen")]
-
-    def __init__(
-        self,
-        error_msg: str,
-        name: str | None = None,
-        id: str | None = None,
-        classes: str | None = None,
-    ) -> None:
-        self.error_msg = error_msg
-        super().__init__(name, id, classes)
-
-    def compose(self) -> ComposeResult:
-        yield Vertical(Label(self.error_msg), Button("OK"))
-
-    def on_button_pressed(self, event: Button.Pressed) -> None:
-        self.app.pop_screen()
 
 
 class TableSelectScreen(Screen):
@@ -102,44 +80,5 @@ class TableSelectScreen(Screen):
         list_view.clear()
 
 
-class RegionSelectScreen(Screen):
-    BINDINGS = [("escape", "app.pop_screen", "Pop screen")]
-
-    class RegionSelected(Message):
-        """"""
-
-        def __init__(self, region: str) -> None:
-            self.region = region
-            super().__init__()
-
-    def compose(self) -> ComposeResult:
-        yield ListView(
-            *[ListItem(Label(region), id=region) for region in get_all_regions()],
-            id="regions",
-        )
-
-    async def on_list_view_selected(self, selected) -> None:
-        self.post_message(self.RegionSelected(selected.item.id))
-        self.app.pop_screen()
 
 
-class ProfileSelectScreen(Screen):
-    BINDINGS = [("escape", "app.pop_screen", "Pop screen")]
-
-    class ProfileSelected(Message):
-        """"""
-
-        def __init__(self, profile: str) -> None:
-            self.profile = profile
-            super().__init__()
-
-    def compose(self) -> ComposeResult:
-        profiles = self.parent.profiles
-        yield ListView(
-            *[ListItem(Label(profile), id=profile) for profile in profiles],
-            id="profiles",
-        )
-
-    async def on_list_view_selected(self, selected) -> None:
-        self.post_message(self.ProfileSelected(selected.item.id))
-        self.app.pop_screen()
